fixed: decimal to float

+++ rest/statistics.py
@@ -429,7 +429,7 @@ class ConsolePerformanceRatingApi(Resource):
             total = clan.num_matches
             wins = ConsoleMatch.objects(win_cond1 | win_cond2).count()
 
-            pr = (sum(strengths) + 400 * (wins - (total - wins))) / total
+            pr = float((sum(strengths) + 400 * (wins - (total - wins))) / total)
 
             return get_response({"pr": round(pr, 2)})
